+++ packages/eyes-sdk-core/lib/EyesBase.js
@@ -14,6 +14,7 @@ const {
   FileDebugScreenshotsProvider,
   NullDebugScreenshotProvider,
   SessionType,
+  GeneralUtils,
 } = require('@applitools/eyes-common')
 
 const {AppOutputProvider} = require('./capture/AppOutputProvider')
@@ -607,6 +608,8 @@ class EyesBase extends EyesAbstract {
 
     // default result
     const validationResult = new ValidationResult()
+
+    await GeneralUtils.sleep(this._configuration.getWaitBeforeScreenshots())
 
     await this.beforeMatchWindow()
     await this._sessionEventHandlers.validationWillStart(this._autSessionId, validationInfo)
@@ -1376,11 +1379,18 @@ class EyesBase extends EyesAbstract {
     let screenshot, screenshotUrl, screenshotBuffer
 
     // START NOTE (amit): the following is something I'm not proud nor confident of. I copied it from EyesSelenium::getScreenshot, and it is to solve https://trello.com/c/O5sTPAU1. This should be rewritten asap.
-    const isMobileDevice = await this._driver.isMobile()
+    let isMobileDevice, originalPosition
     const positionProvider = this.getPositionProvider()
-    let originalPosition
-    if (!isMobileDevice && positionProvider) {
-      originalPosition = await positionProvider.getState()
+    if (this._driver) {
+      if (typeof this._driver.isMobile === 'function') {
+        isMobileDevice = await this._driver.isMobile()
+      } else if (typeof this._driver.isMobile === 'object') {
+        isMobileDevice = await this._driver.isMobile
+      }
+
+      if (!isMobileDevice && positionProvider) {
+        originalPosition = await positionProvider.getState()
+      }
     }
     // END NOTE
 
@@ -1538,7 +1548,7 @@ class EyesBase extends EyesAbstract {
    * @param {RectangleSize} size - The required viewport size.
    * @return {Promise}
    */
-  async setViewportSize(size) {
+  async setViewportSize(_size) {
     // eslint-disable-line no-unused-vars
     throw new TypeError('The method is not implemented!')
   }