+++ packages/eyes-testcafe/lib/fluent/FloatingRegionByElement.js
@@ -30,8 +30,7 @@ class FloatingRegionByElement extends GetFloatingRegion {
    * @param {EyesScreenshot} screenshot
    * @return {Promise<FloatingMatchSettings[]>}
    */
-  async getRegion(eyes, screenshot) {
-    // eslint-disable-line no-unused-vars
+  async getRegion(_eyes, screenshot) {
     const rect = await this._element.getRect()
     const lTag = screenshot.convertLocation(
       new Location(rect),