+++ packages/eyes-testcafe/scripts/play/play.testcafe.js
@@ -3,8 +3,14 @@
 
 'use strict'
 
-const {Configuration, StitchMode} = require('@applitools/eyes-common')
-const {Eyes, Target, ConsoleLogHandler} = require('../..')
+const {
+  Eyes,
+  Target,
+  ConsoleLogHandler,
+  FileDebugScreenshotsProvider,
+  Configuration,
+  StitchMode,
+} = require('../../index')
 
 /*
  * Play with configuration and test :
@@ -13,36 +19,50 @@ const {Eyes, Target, ConsoleLogHandler} = require('../..')
 const eyes = new Eyes()
 const configuration = new Configuration({
   stitchMode: StitchMode.CSS,
-  stitchOverlap: 56,
-  viewportSize: {width: 1024, height: 768},
+  // stitchOverlap: 56,
+  viewportSize: {width: 600, height: 500},
 })
+
+// const debugHandler = new FileDebugScreenshotsProvider()
+// debugHandler.setPath('./screenshots')
+// eyes.setDebugScreenshotsProvider(debugHandler)
 eyes.setConfiguration(configuration)
 
 if (process.env.APPLITOOLS_SHOW_LOGS || process.env.LIVE) {
   eyes.setLogHandler(new ConsoleLogHandler(true))
 }
 
-fixture`Play`.page`https://www.apple.com/apple-watch-series-3/`
+fixture`Play`.page`http://localhost:8080/full-page.html`
+test('Play', async t => {
+  await eyes.open(t, 'Play Testcafe', 'play testcafe')
+  // await eyes._scanPage()
+  await eyes.check('page play', Target.window().fully())
+  const result = await eyes.close(false)
+  console.log('Play result', result)
+})
 
 // test('Play', async t => {
-//   await eyes.open(t, 'Play Testcafe', 'play testcafe')
-//   // await eyes._scanPage()
-//   await eyes.check('page play', Target.window().fully())
-//   const result = await eyes.close(false)
-//   console.log('Play result', result)
+//   // const captureFrameAndPollForIE = require('../../dist/captureFrameAndPollForIE')
+//   // const {TestCafeExecutor} = require('../../lib/TestCafeExecutor')
+//   // const ex = new TestCafeExecutor(t)
+//   // console.log('XXXXXXXXX: captureFrameAndPollForIE', captureFrameAndPollForIE.toString())
+//   // const r = await ex.executeScript(captureFrameAndPollForIE)
+//   // const r2 = await ex.executeScript(captureFrameAndPollForIE)
+//   // console.log('XXXXXXXXX: r', r2)
+
+//   await t.resizeWindow(600, 500)
+//   const transformElement = require('../../lib/positioning/transformElement')
+//   await t.eval(transformElement, {
+//     dependencies: {
+//       element: () => document.documentElement,
+//       transformLeft: -600,
+//       transformTop: 0,
+//       markRightMargin: 10,
+//       originalTransform: undefined,
+//     },
+//   })
+
+//   const name = new Date().toISOString().replace(/:/g, '_')
+//   const image = await t.takeScreenshot(`./render-${name}.png`)
+//   console.log('XXXXXXXXX: image', image)
 // })
-
-test('Play', async t => {
-  // const captureFrameAndPollForIE = require('../../dist/captureFrameAndPollForIE')
-  // const {TestCafeExecutor} = require('../../lib/TestCafeExecutor')
-  // const ex = new TestCafeExecutor(t)
-  // console.log('XXXXXXXXX: captureFrameAndPollForIE', captureFrameAndPollForIE.toString())
-  // const r = await ex.executeScript(captureFrameAndPollForIE)
-  // const r2 = await ex.executeScript(captureFrameAndPollForIE)
-  // console.log('XXXXXXXXX: r', r2)
-
-  await t.resizeWindow(600, 500)
-  const name = new Date().toISOString().replace(/:/g, '_')
-  const image = await t.takeScreenshot(`./render-${name}.png`)
-  console.log('XXXXXXXXX: image', image)
-})