+++ packages/eyes-testcafe/scripts/render.js
@@ -33,12 +33,13 @@ createTestCafe(null, 1339)
         'chrome:headless',
         'browserstack:safari@13.0:OS X Catalina',
         'browserstack:safari@12.1:OS X Mojave',
-        'browserstack:iPhone XS',
-        'browserstack:iPhone X',
-        'browserstack:iPhone 8',
-        'browserstack:iPad Pro 12.9 2018',
         'browserstack:edge',
         'browserstack:ie',
+        'saucelabs:Safari@13.0:macOS 10.13',
+        'browserstack:iPhone XS',
+        'saucelabs:iPhone XS Simulator@13.0',
+        'saucelabs:iPhone Simulator@13.0',
+        'browserstack:iPad Pro 12.9 2018',
         'browserstack:chrome',
       ][0]
     } else {