+++ packages/eyes-testcafe/scripts/render.testcafe.js
@@ -7,7 +7,7 @@ const eyes = new Eyes()
 const configuration = new Configuration({
   stitchMode: StitchMode.CSS,
   // stitchOverlap: 56,
-  viewportSize: {width: 600, height: 500},
+  viewportSize: {width: 1024, height: 768},
 })
 eyes.setConfiguration(configuration)