+++ packages/eyes-testcafe/test/e2e/testcafe/browsers/helloworld-fullpage.edge.testcafe.js
@@ -2,21 +2,9 @@
 
 'use strict'
 
-const {Configuration, StitchMode} = require('@applitools/eyes-common')
-const {Eyes, Target, ConsoleLogHandler} = require('../../../..')
+const {Configuration, StitchMode, Eyes, Target, ConsoleLogHandler} = require('../../../../index')
 
-fixture`Hello world full page Edge`.page`https://www.applitools.com/helloworld`
-
-test('helloworld full page Edge', async t => {
-  const eyes = new Eyes()
-  eyes.setConfiguration(new Configuration({viewportSize: {width: 600, height: 500}}))
-  if (process.env.APPLITOOLS_SHOW_LOGS || process.env.APPLITOOLS_DEBUG_TEST) {
-    eyes.setLogHandler(new ConsoleLogHandler(true))
-  }
-  await eyes.open(t, 'Applitools helloworld', 'eyes-testcafe e2e - fullpage')
-  await eyes.check('some tag', Target.window().fully())
-  await eyes.close()
-})
+fixture`full page Edge`.page`http://localhost:5556/full-page.html`
 
 test('full page css stiching Edge', async t => {
   const eyes = new Eyes()
@@ -26,7 +14,18 @@ test('full page css stiching Edge', async t => {
   if (process.env.APPLITOOLS_SHOW_LOGS || process.env.APPLITOOLS_DEBUG_TEST) {
     eyes.setLogHandler(new ConsoleLogHandler(true))
   }
-  await eyes.open(t, 'Applitools helloworld', 'eyes-testcafe e2e - fullpage')
+  await eyes.open(t, 'Applitools full page Edge', 'eyes-testcafe e2e - fullpage')
+  await eyes.check('some tag', Target.window().fully())
+  await eyes.close()
+})
+
+test('full page Edge', async t => {
+  const eyes = new Eyes()
+  eyes.setConfiguration(new Configuration({viewportSize: {width: 600, height: 500}}))
+  if (process.env.APPLITOOLS_SHOW_LOGS || process.env.APPLITOOLS_DEBUG_TEST) {
+    eyes.setLogHandler(new ConsoleLogHandler(true))
+  }
+  await eyes.open(t, 'Applitools full page Edge', 'eyes-testcafe e2e - fullpage')
   await eyes.check('some tag', Target.window().fully())
   await eyes.close()
 })