+++ packages/visual-grid-client/src/sdk/openEyes.js
@@ -6,6 +6,7 @@ const makeCheckWindow = require('./checkWindow')
 const makeAbort = require('./makeAbort')
 const makeClose = require('./makeClose')
 const translateBrowserNameVersion = require('./translateBrowserNameVersion')
+const isEmulation = require('./isEmulation')
 
 const {
   initWrappers,
@@ -15,22 +16,26 @@ const {
   apiKeyFailMsg,
 } = require('./wrapperUtils')
 
-const SUPPORTED_BROWSERS = [
-  'firefox',
-  'ie10',
-  'ie11',
-  'edge',
-  'chrome',
-  'ie',
-  'safari',
-  'chrome-canary',
-  'chrome-1',
-  'chrome-2',
-  'firefox-1',
-  'firefox-2',
-  'safari-1',
-  'safari-2',
-]
+// This is a map from the value we get from the user to the value we send to the visual grid
+// user --> VG
+const SUPPORTED_BROWSERS = {
+  chrome: 'chrome',
+  'chrome-canary': 'chrome-canary',
+  firefox: 'firefox',
+  ie10: 'ie10',
+  ie11: 'ie11',
+  edge: 'edge',
+  ie: 'ie',
+  safari: 'safari',
+  [translateBrowserNameVersion('chrome-1')]: 'chrome-1',
+  [translateBrowserNameVersion('chrome-2')]: 'chrome-2',
+  [translateBrowserNameVersion('firefox-1')]: 'firefox-1',
+  [translateBrowserNameVersion('firefox-2')]: 'firefox-2',
+  [translateBrowserNameVersion('safari-1')]: 'safari-1',
+  [translateBrowserNameVersion('safari-2')]: 'safari-2',
+}
+const SUPPORTED_BROWSER_KEYS = Object.keys(SUPPORTED_BROWSERS)
+const SUPPORTED_BROWSER_KEYS_STR = `\n* ${SUPPORTED_BROWSER_KEYS.join('\n* ')}\n`
 
 function makeOpenEyes({
   appName: _appName,
@@ -139,17 +144,18 @@ function makeOpenEyes({
     }
 
     const browsersArray = Array.isArray(browser) ? browser : [browser]
-    const browsers = browsersArray.map(browser => ({
-      ...browser,
-      name: translateBrowserNameVersion(browser.name),
-    }))
-    const browserError = browsers.length
-      ? browsers.map(getBrowserError).find(Boolean)
+    const browserError = browsersArray.length
+      ? browsersArray.map(getBrowserError).find(Boolean)
       : getBrowserError()
     if (browserError) {
       console.log('\x1b[31m', `\nInvalid browser: ${browserError}\n`)
       throw new Error(browserError)
     }
+
+    const browsers = browsersArray.map(browser => ({
+      ...browser,
+      name: SUPPORTED_BROWSERS[browser.name] || browser.name,
+    }))
 
     ;({batchSequence, baselineBranch, parentBranch, branch, batchNotify} = backwardCompatible(
       [{batchSequenceName}, {batchSequence}],
@@ -300,12 +306,19 @@ function makeOpenEyes({
       if (!browser) {
         return 'invalid browser configuration provided.'
       }
-      if (browser.name && !SUPPORTED_BROWSERS.includes(browser.name)) {
-        return `browser name should be one of the following 'chrome', 'firefox', 'safari', 'ie10', 'ie11' or 'edge' but received '${browser.name}'.`
+      if (browser.name && !SUPPORTED_BROWSER_KEYS.includes(browser.name)) {
+        return `browser name should be one of the following:${SUPPORTED_BROWSER_KEYS_STR}\nReceived: '${browser.name}'.`
       }
       if (browser.name && !browser.deviceName && (!browser.height || !browser.width)) {
         return `browser '${browser.name}' should include 'height' and 'width' parameters.`
       }
+      if (isEmulation(browser) && !isSupportsDeviceEmulation(browser.name)) {
+        return `browser '${browser.name}' does not support mobile device emulation. Please remove 'mobile:true' or 'deviceName' from the browser configuration`
+      }
+    }
+
+    function isSupportsDeviceEmulation(browserName) {
+      return !browserName || /^chrome/.test(browserName)
     }
   }
 }