+++ packages/visual-grid-client/src/sdk/translateBrowserNameVersion.js
@@ -1,15 +1,15 @@
 'use strict'
 
-const ONE_VERSION_REGEX = /^(chrome|firefox|safari)\-one\-version\-back$/
-const TWO_VERSIONS_REGEX = /^(chrome|firefox|safari)\-two\-versions\-back$/
+const ONE_VERSION_REGEX = /^(chrome|firefox|safari)\-1$/
+const TWO_VERSIONS_REGEX = /^(chrome|firefox|safari)\-2$/
 
 function translateBrowserNameVersion(browserName) {
   if (ONE_VERSION_REGEX.test(browserName)) {
-    return browserName.replace('one-version-back', '1')
+    return browserName.replace('1', 'one-version-back')
   }
 
   if (TWO_VERSIONS_REGEX.test(browserName)) {
-    return browserName.replace('two-versions-back', '2')
+    return browserName.replace('2', 'two-versions-back')
   }
 
   return browserName